Fix compiling x509_tbscert_strip_ct_ext with OpenSSL 1.1.0
As the X509 structure is opaque in OpenSSL 1.1.0, it's not possible to
use the cert_info pointer directly.

This commit replaces the code for setting the modified flag and the
call to i2d_X509_CINF with i2d_re_X509_tbs, which does exactly the same
thing but takes a pointer to the X509 structure instead of the X509_CINF
structure.

An implementation of i2d_re_X509_tbs is included for compatibility with
OpenSSL 1.0.2 and below.